login test convert to junit
